Remove target for Home page link on spaghetti page
Navigating to the homepage from the spaghetti recipe page without this change would open the homepage on a new tab.
If users use the links to switch between the home page and recipe pages, it will result in duplicate tabs of the same webpage.
Making the homepage open on the same tab saves space on the tab bar.